forgot to transcribe pairing board numbers
The board numbers for each round are not really important to document.  (They're only relevant to when we need to find our seats and meet our opponent at the right table.)  However, the PGN specification does include a `Board` tag, and because I had written the trivial numbers down on my score sheet, it just eases my conscience to transcribe them here to the electronic copy before finally throwing the paper clutter away.